# infrastructure/rest/controllers/auth/Structures.go
package auth

type LoginRequest struct {
	Email    string `json:"email" example:"mail@mail.com" gorm:"unique" binding:"required"`
	Password string `json:"password" example:"Password123" binding:"required"`
}